Started refactoring eventhubs to enable centralized retries
