+++ sdk/core/azure_core_amqp/src/messaging.rs
@@ -342,7 +342,7 @@ impl From<AmqpList> for AmqpTarget {
                 let dynamic_node_properties: AmqpOrderedMap<String, AmqpValue> =
                     dynamic_node_properties
                         .iter()
-                        .map(|(k, v)| (k.into(), v))
+                        .map(|(k, v)| (k.into(), v.clone()))
                         .collect();
                 builder = builder.with_dynamic_node_properties(dynamic_node_properties);
             }
@@ -488,7 +488,7 @@ impl From<AmqpList> for AmqpSource {
                 let dynamic_node_properties: AmqpOrderedMap<AmqpSymbol, AmqpValue> =
                     dynamic_node_properties
                         .iter()
-                        .map(|(k, v)| (k.into(), v))
+                        .map(|(k, v)| (k.into(), v.clone()))
                         .collect();
                 builder = builder.with_dynamic_node_properties(dynamic_node_properties);
             }
@@ -501,7 +501,7 @@ impl From<AmqpList> for AmqpSource {
         if field_count >= 8 {
             if let Some(AmqpValue::Map(filter)) = list.0.get(7) {
                 let filter: AmqpOrderedMap<AmqpSymbol, AmqpValue> =
-                    filter.iter().map(|(k, v)| (k.into(), v)).collect();
+                    filter.iter().map(|(k, v)| (k.into(), v.clone())).collect();
                 builder = builder.with_filter(filter);
             }
         }
