+++ sdk/eventhubs/azure_messaging_eventhubs/src/common/management.rs
@@ -1,23 +1,17 @@
 // Copyright (c) Microsoft Corporation. All Rights reserved
 // Licensed under the MIT license.
 
-use super::{retry_azure_operation, RetryOptions};
+use super::recoverable_connection::RecoverableConnection;
 use crate::{
     error::{ErrorKind, EventHubsError},
     models::{EventHubPartitionProperties, EventHubProperties},
 };
-use azure_core::error::{ErrorKind as AzureErrorKind, Result};
-use azure_core_amqp::{
-    error::{AmqpErrorCondition, AmqpErrorKind},
-    AmqpError, AmqpManagement, AmqpManagementApis, AmqpOrderedMap, AmqpSimpleValue, AmqpTimestamp,
-    AmqpValue,
-};
-use std::{error::Error, time::SystemTime};
-use tracing::{debug, warn};
+use azure_core::error::Result;
+use azure_core_amqp::{AmqpOrderedMap, AmqpSimpleValue, AmqpTimestamp, AmqpValue};
+use std::{sync::Arc, time::SystemTime};
 
 pub(crate) struct ManagementInstance {
-    pub management: AmqpManagement,
-    retry_options: RetryOptions,
+    connection: Arc<RecoverableConnection>,
 }
 
 const EVENTHUB_ENTITY_TYPE: &str = "com.microsoft:eventhub";
@@ -40,87 +34,19 @@ const EVENTHUB_PARTITION_PROPERTIES_LAST_ENQUEUED_TIME_UTC: &str = "last_enqueue
 const EVENTHUB_PARTITION_PROPERTIES_IS_EMPTY: &str = "is_partition_empty";
 
 impl ManagementInstance {
-    pub fn new(management: AmqpManagement, retry_options: RetryOptions) -> Self {
-        Self {
-            management,
-            retry_options,
-        }
-    }
-
-    fn should_retry_management_response(e: &azure_core::Error) -> bool {
-        match e.kind() {
-            AzureErrorKind::Amqp => {
-                warn!("Amqp operation failed: {}", e.source().unwrap());
-                if let Some(e) = e.source() {
-                    debug!("Error: {}", e);
-
-                    if let Some(amqp_error) = e.downcast_ref::<Box<AmqpError>>() {
-                        Self::should_retry_amqp_error(amqp_error)
-                    } else if let Some(amqp_error) = e.downcast_ref::<AmqpError>() {
-                        Self::should_retry_amqp_error(amqp_error)
-                    } else {
-                        debug!("Non AMQP error: {}", e);
-                        false
-                    }
-                } else {
-                    debug!("No source error found");
-                    false
-                }
-            }
-            _ => {
-                debug!("Non AMQP error: {}", e);
-                false
-            }
-        }
-    }
-
-    fn should_retry_amqp_error(amqp_error: &AmqpError) -> bool {
-        match amqp_error.kind() {
-            AmqpErrorKind::ManagementStatusCode(code, _) => {
-                debug!("Management operation error: {}", code);
-                match code {
-                    // Retry on 408 (Request Timeout) and 429 (Too Many Requests)
-                    azure_core::http::StatusCode::RequestTimeout
-                    | azure_core::http::StatusCode::TooManyRequests
-                    | azure_core::http::StatusCode::InternalServerError
-                    | azure_core::http::StatusCode::BadGateway
-                    | azure_core::http::StatusCode::ServiceUnavailable
-                    | azure_core::http::StatusCode::GatewayTimeout => true,
-                    _ => false,
-                }
-            }
-            AmqpErrorKind::AmqpDescribedError(described_error) => {
-                debug!("AMQP described error: {:?}", described_error);
-                matches!(
-                    described_error.condition(),
-                    AmqpErrorCondition::ResourceLimitExceeded
-                        | AmqpErrorCondition::ConnectionFramingError
-                        | AmqpErrorCondition::LinkStolen
-                )
-            }
-            _ => {
-                debug!("Other AMQP error: {}", amqp_error);
-                false
-            }
-        }
+    pub fn new(connection: Arc<RecoverableConnection>) -> Arc<Self> {
+        Arc::new(Self { connection })
     }
 
     pub async fn get_eventhub_properties(&self, eventhub: &str) -> Result<EventHubProperties> {
-        let response = retry_azure_operation(
-            || async move {
-                let mut application_properties: AmqpOrderedMap<String, AmqpSimpleValue> =
-                    AmqpOrderedMap::new();
-                application_properties.insert(EVENTHUB_PROPERTY_NAME.to_string(), eventhub.into());
-                let response = self
-                    .management
-                    .call(EVENTHUB_ENTITY_TYPE.to_string(), application_properties)
-                    .await?;
-                Ok(response)
-            },
-            &self.retry_options,
-            Some(Self::should_retry_management_response),
-        )
-        .await?;
+        let mut application_properties: AmqpOrderedMap<String, AmqpSimpleValue> =
+            AmqpOrderedMap::new();
+        application_properties.insert(EVENTHUB_PROPERTY_NAME.to_string(), eventhub.into());
+
+        let response = self
+            .connection
+            .call_amqp_management(EVENTHUB_ENTITY_TYPE.to_string(), application_properties)
+            .await?;
 
         if !response.contains_key(EVENTHUB_PROPERTY_PARTITION_COUNT) {
             return Err(EventHubsError::from(ErrorKind::InvalidManagementResponse).into());
@@ -163,22 +89,15 @@ impl ManagementInstance {
         eventhub: &str,
         partition_id: &str,
     ) -> Result<EventHubPartitionProperties> {
-        let response = retry_azure_operation(
-            || async move {
-                let mut application_properties: AmqpOrderedMap<String, AmqpSimpleValue> =
-                    AmqpOrderedMap::new();
-                application_properties.insert(EVENTHUB_PROPERTY_NAME.to_string(), eventhub.into());
-                application_properties
-                    .insert(EVENTHUB_PROPERTY_PARTITION.to_string(), partition_id.into());
+        let mut application_properties: AmqpOrderedMap<String, AmqpSimpleValue> =
+            AmqpOrderedMap::new();
+        application_properties.insert(EVENTHUB_PROPERTY_NAME.to_string(), eventhub.into());
+        application_properties.insert(EVENTHUB_PROPERTY_PARTITION.to_string(), partition_id.into());
 
-                self.management
-                    .call(PARTITION_ENTITY_TYPE.to_string(), application_properties)
-                    .await
-            },
-            &self.retry_options,
-            Some(Self::should_retry_management_response),
-        )
-        .await?;
+        let response = self
+            .connection
+            .call_amqp_management(PARTITION_ENTITY_TYPE.to_string(), application_properties)
+            .await?;
 
         // Look for the required response properties
         if !response.contains_key(EVENTHUB_PARTITION_PROPERTIES_TYPE)
@@ -221,105 +140,5 @@ impl ManagementInstance {
                 .ok_or_else(|| EventHubsError::from(ErrorKind::InvalidManagementResponse))?
                 .into(),
         })
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use azure_core_amqp::AmqpError;
-
-    use crate::consumer;
-
-    use super::*;
-
-    #[test]
-    fn should_retry_management_response() {
-        consumer::tests::setup();
-
-        {
-            let error: azure_core::Error = AmqpError::new_management_error(
-                azure_core::http::StatusCode::TooManyRequests,
-                Some("Too many requests!".into()),
-            )
-            .into();
-
-            assert!(ManagementInstance::should_retry_management_response(&error));
-        }
-        {
-            let error: azure_core::Error = AmqpError::new_management_error(
-                azure_core::http::StatusCode::SwitchingProtocols,
-                Some("Switcheroo".into()),
-            )
-            .into();
-            assert!(!ManagementInstance::should_retry_management_response(
-                &error
-            ));
-        }
-        // Verify that an explicitly boxed error is handled correctly
-        {
-            let error = azure_core::Error::new(
-                AzureErrorKind::Amqp,
-                Box::new(AmqpError::new_management_error(
-                    azure_core::http::StatusCode::TooManyRequests,
-                    Some("Too many requests!".into()),
-                )),
-            );
-            assert!(ManagementInstance::should_retry_management_response(&error));
-        }
-
-        {
-            let error: azure_core::Error = AmqpError::new_management_error(
-                azure_core::http::StatusCode::BadGateway,
-                Some("Bad Gateway".into()),
-            )
-            .into();
-            assert!(ManagementInstance::should_retry_management_response(&error));
-        }
-        {
-            let error: azure_core::Error = AmqpError::new_management_error(
-                azure_core::http::StatusCode::RequestTimeout,
-                Some("Request Timeout".into()),
-            )
-            .into();
-            assert!(ManagementInstance::should_retry_management_response(&error));
-        }
-        {
-            let error: azure_core::Error = AmqpError::new_management_error(
-                azure_core::http::StatusCode::InternalServerError,
-                Some("Internal Server Error".into()),
-            )
-            .into();
-            assert!(ManagementInstance::should_retry_management_response(&error));
-            {
-                let error: azure_core::Error =
-                    EventHubsError::from(ErrorKind::InvalidManagementResponse).into();
-                assert!(!ManagementInstance::should_retry_management_response(
-                    &error
-                ));
-            }
-
-            {
-                let error: azure_core::Error = AmqpError::new_described_error(
-                    AmqpErrorCondition::ResourceLimitExceeded,
-                    Some("Resource Limit Exceeded".into()),
-                    Default::default(),
-                )
-                .into();
-
-                assert!(ManagementInstance::should_retry_management_response(&error));
-            }
-            {
-                let error: azure_core::Error = AmqpError::new_described_error(
-                    AmqpErrorCondition::IllegalState,
-                    Some("Illegal State".into()),
-                    Default::default(),
-                )
-                .into();
-
-                assert!(!ManagementInstance::should_retry_management_response(
-                    &error
-                ));
-            }
-        }
     }
 }
