+++ sdk/eventhubs/azure_messaging_eventhubs/src/models/event_data.rs
@@ -198,7 +198,7 @@ impl ReceivedEventData {
 
             for (key, value) in annotations.0.iter() {
                 if let AmqpAnnotationKey::Symbol(symbol) = key {
-                    if symbol == ENQUEUED_TIME_UTC {
+                    if *symbol == ENQUEUED_TIME_UTC {
                         if let AmqpValue::TimeStamp(timestamp) = value {
                             return timestamp.0;
                         }
@@ -216,9 +216,9 @@ impl ReceivedEventData {
             let annotations = self.message.message_annotations()?;
             for (key, value) in annotations.0.iter() {
                 if let AmqpAnnotationKey::Symbol(symbol) = key {
-                    if symbol == OFFSET {
+                    if *symbol == OFFSET {
                         if let AmqpValue::String(offset_value) = value {
-                            return Some(offset_value);
+                            return Some(offset_value.clone());
                         }
                     }
                 }
@@ -233,9 +233,9 @@ impl ReceivedEventData {
             let annotations = self.message.message_annotations()?;
             for (key, value) in annotations.0.iter() {
                 if let AmqpAnnotationKey::Symbol(symbol) = key {
-                    if symbol == SEQUENCE_NUMBER {
+                    if *symbol == SEQUENCE_NUMBER {
                         if let AmqpValue::Long(sequence_number_value) = value {
-                            return Some(sequence_number_value);
+                            return Some(*sequence_number_value);
                         }
                     }
                 }
@@ -252,9 +252,9 @@ impl ReceivedEventData {
             let annotations = self.message.message_annotations()?;
             for (key, value) in annotations.0.iter() {
                 if let AmqpAnnotationKey::Symbol(symbol) = key {
-                    if symbol == PARTITION_KEY {
+                    if *symbol == PARTITION_KEY {
                         if let AmqpValue::String(partition_key_value) = value {
-                            return Some(partition_key_value);
+                            return Some(partition_key_value.clone());
                         }
                     }
                 }
@@ -273,10 +273,10 @@ impl ReceivedEventData {
             if let Some(annotations) = self.message.message_annotations() {
                 for (key, value) in annotations.0.iter() {
                     if let AmqpAnnotationKey::Symbol(symbol) = key {
-                        if symbol != ENQUEUED_TIME_UTC
-                            && symbol != OFFSET
-                            && symbol != SEQUENCE_NUMBER
-                            && symbol != PARTITION_KEY
+                        if *symbol != ENQUEUED_TIME_UTC
+                            && *symbol != OFFSET
+                            && *symbol != SEQUENCE_NUMBER
+                            && *symbol != PARTITION_KEY
                         {
                             system_properties.insert(symbol.0.clone(), value.clone());
                         }
